Move em params to subdirectory
