+++ src/celeritas/em/params/UrbanMscParams.hh
@@ -1,9 +1,9 @@
 //----------------------------------*-C++-*----------------------------------//
-// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
+// Copyright 2023-2024 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/em/WentzelVIMscParams.hh
+//! \file celeritas/em/params/UrbanMscParams.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
@@ -11,7 +11,7 @@
 
 #include "corecel/data/CollectionMirror.hh"
 #include "corecel/data/ParamsDataInterface.hh"
-#include "celeritas/em/data/WentzelVIMscData.hh"
+#include "celeritas/em/data/UrbanMscData.hh"
 
 namespace celeritas
 {
@@ -24,11 +24,11 @@ struct ImportMscModel;
 
 //---------------------------------------------------------------------------//
 /*!
- * Construct and store data for Wentzel VI multiple scattering.
+ * Construct and store data for Urban multiple scattering.
  *
  * Multiple scattering is used by the along-step kernel(s).
  */
-class WentzelVIMscParams final : public ParamsDataInterface<WentzelVIMscData>
+class UrbanMscParams final : public ParamsDataInterface<UrbanMscData>
 {
   public:
     //!@{
@@ -38,25 +38,31 @@ class WentzelVIMscParams final : public ParamsDataInterface<WentzelVIMscData>
 
   public:
     // Construct if MSC process data is present, else return nullptr
-    static std::shared_ptr<WentzelVIMscParams>
+    static std::shared_ptr<UrbanMscParams>
     from_import(ParticleParams const& particles,
                 MaterialParams const& materials,
                 ImportData const& data);
 
     // Construct from process data
-    WentzelVIMscParams(ParticleParams const& particles,
-                       MaterialParams const& materials,
-                       VecImportMscModel const& mdata);
+    UrbanMscParams(ParticleParams const& particles,
+                   MaterialParams const& materials,
+                   VecImportMscModel const& mdata);
 
-    //! Access Wentzel VI data on the host
+    // TODO: possible "applicability" interface used for constructing
+    // along-step kernels?
+
+    //! Access UrbanMsc data on the host
     HostRef const& host_ref() const final { return data_.host_ref(); }
 
-    //! Access Wentzel VI data on the device
+    //! Access UrbanMsc data on the device
     DeviceRef const& device_ref() const final { return data_.device_ref(); }
 
   private:
     // Host/device storage and reference
-    CollectionMirror<WentzelVIMscData> data_;
+    CollectionMirror<UrbanMscData> data_;
+
+    static UrbanMscMaterialData
+    calc_material_data(MaterialView const& material_view);
 };
 
 //---------------------------------------------------------------------------//
