+++ src/celeritas/em/params/WentzelVIMscParams.hh
@@ -1,9 +1,9 @@
 //----------------------------------*-C++-*----------------------------------//
-// Copyright 2023-2024 UT-Battelle, LLC, and other Celeritas developers.
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/em/UrbanMscParams.hh
+//! \file celeritas/em/params/WentzelVIMscParams.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
@@ -11,8 +11,7 @@
 
 #include "corecel/data/CollectionMirror.hh"
 #include "corecel/data/ParamsDataInterface.hh"
-
-#include "data/UrbanMscData.hh"
+#include "celeritas/em/data/WentzelVIMscData.hh"
 
 namespace celeritas
 {
@@ -25,11 +24,11 @@ struct ImportMscModel;
 
 //---------------------------------------------------------------------------//
 /*!
- * Construct and store data for Urban multiple scattering.
+ * Construct and store data for Wentzel VI multiple scattering.
  *
  * Multiple scattering is used by the along-step kernel(s).
  */
-class UrbanMscParams final : public ParamsDataInterface<UrbanMscData>
+class WentzelVIMscParams final : public ParamsDataInterface<WentzelVIMscData>
 {
   public:
     //!@{
@@ -39,31 +38,25 @@ class UrbanMscParams final : public ParamsDataInterface<UrbanMscData>
 
   public:
     // Construct if MSC process data is present, else return nullptr
-    static std::shared_ptr<UrbanMscParams>
+    static std::shared_ptr<WentzelVIMscParams>
     from_import(ParticleParams const& particles,
                 MaterialParams const& materials,
                 ImportData const& data);
 
     // Construct from process data
-    UrbanMscParams(ParticleParams const& particles,
-                   MaterialParams const& materials,
-                   VecImportMscModel const& mdata);
+    WentzelVIMscParams(ParticleParams const& particles,
+                       MaterialParams const& materials,
+                       VecImportMscModel const& mdata);
 
-    // TODO: possible "applicability" interface used for constructing
-    // along-step kernels?
-
-    //! Access UrbanMsc data on the host
+    //! Access Wentzel VI data on the host
     HostRef const& host_ref() const final { return data_.host_ref(); }
 
-    //! Access UrbanMsc data on the device
+    //! Access Wentzel VI data on the device
     DeviceRef const& device_ref() const final { return data_.device_ref(); }
 
   private:
     // Host/device storage and reference
-    CollectionMirror<UrbanMscData> data_;
-
-    static UrbanMscMaterialData
-    calc_material_data(MaterialView const& material_view);
+    CollectionMirror<WentzelVIMscData> data_;
 };
 
 //---------------------------------------------------------------------------//
