+++ test/celeritas/em/Fluctuation.test.cc
@@ -8,12 +8,12 @@
 #include "corecel/data/CollectionStateStore.hh"
 #include "celeritas/MockTestBase.hh"
 #include "celeritas/Quantities.hh"
-#include "celeritas/em/FluctuationParams.hh"
 #include "celeritas/em/distribution/EnergyLossDeltaDistribution.hh"
 #include "celeritas/em/distribution/EnergyLossGammaDistribution.hh"
 #include "celeritas/em/distribution/EnergyLossGaussianDistribution.hh"
 #include "celeritas/em/distribution/EnergyLossHelper.hh"
 #include "celeritas/em/distribution/EnergyLossUrbanDistribution.hh"
+#include "celeritas/em/params/FluctuationParams.hh"
 #include "celeritas/mat/MaterialParams.hh"
 #include "celeritas/phys/CutoffParams.hh"
 #include "celeritas/phys/ParticleParams.hh"
