+++ test/celeritas/em/LivermorePE.test.cc
@@ -14,10 +14,10 @@
 #include "corecel/math/ArrayUtils.hh"
 #include "corecel/sys/Device.hh"
 #include "celeritas/Quantities.hh"
-#include "celeritas/em/AtomicRelaxationParams.hh"
 #include "celeritas/em/detail/Utils.hh"
 #include "celeritas/em/interactor/LivermorePEInteractor.hh"
 #include "celeritas/em/model/LivermorePEModel.hh"
+#include "celeritas/em/params/AtomicRelaxationParams.hh"
 #include "celeritas/em/xs/LivermorePEMicroXsCalculator.hh"
 #include "celeritas/grid/ValueGridBuilder.hh"
 #include "celeritas/grid/ValueGridInserter.hh"
