+++ test/celeritas/em/UrbanMsc.test.cc
@@ -14,12 +14,12 @@
 #include "corecel/grid/Interpolator.hh"
 #include "geocel/UnitUtils.hh"
 #include "celeritas/RootTestBase.hh"
-#include "celeritas/em/UrbanMscParams.hh"
 #include "celeritas/em/msc/detail/MscStepFromGeo.hh"
 #include "celeritas/em/msc/detail/MscStepToGeo.hh"
 #include "celeritas/em/msc/detail/UrbanMscMinimalStepLimit.hh"
 #include "celeritas/em/msc/detail/UrbanMscSafetyStepLimit.hh"
 #include "celeritas/em/msc/detail/UrbanMscScatter.hh"
+#include "celeritas/em/params/UrbanMscParams.hh"
 #include "celeritas/geo/GeoData.hh"
 #include "celeritas/geo/GeoParams.hh"
 #include "celeritas/geo/GeoTrackView.hh"
