+++ test/celeritas/user/Diagnostic.test.cc
@@ -8,7 +8,7 @@
 #include "corecel/cont/Span.hh"
 #include "corecel/io/StringUtils.hh"
 #include "geocel/UnitUtils.hh"
-#include "celeritas/em/UrbanMscParams.hh"
+#include "celeritas/em/params/UrbanMscParams.hh"
 #include "celeritas/ext/GeantPhysicsOptions.hh"
 #include "celeritas/global/ActionRegistry.hh"
 #include "celeritas/global/Stepper.hh"
